pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["lifehackett <chackett@dustidentity.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests-oidc = "^0.1.4"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
